ref: move vesting logic to its own function
