+++ src/components/claiming/Vesting.vue
@@ -8,14 +8,14 @@
     title="Vesting"
   >
     <p>You will unlock the following amount of wits</p>
-    <p v-for="step in steps" :key="step.date">{{ step.date }} - {{ step.amount }} wits</p>
+    <p v-for="step in vesting" :key="step.date">{{ step.date }} - {{ step.amount }} wits</p>
   </NavigationCard>
 </template>
 
 <script>
 import NavigationCard from '@/components/card/NavigationCard'
 import { mapState } from 'vuex'
-import { changeDateFormat } from '@/utils'
+import { calculateVesting } from '@/utils'
 export default {
   name: 'Vesting',
   components: {
@@ -23,36 +23,15 @@ export default {
   },
   computed: {
     ...mapState({
-      cliff: state => state.wallet.claimingFileInfo.info.data.vesting.cliff,
-      delay: state => state.wallet.claimingFileInfo.info.data.vesting.delay,
-      installmentLength: state => state.wallet.claimingFileInfo.info.data.vesting.installmentLength,
-      installmentWits: state => state.wallet.claimingFileInfo.info.data.vesting.installmentWits,
+      vestingInfo: state => state.wallet.claimingFileInfo.info.data.vesting,
       genesisDate: state => state.wallet.claimingFileInfo.info.data.genesisDate,
       amount: state => state.wallet.claimingFileInfo.amount,
     }),
 
-    steps() {
-      const numberOfSteps = Math.ceil(this.amount / this.installmentWits)
-      let amount = this.amount
-      const steps = Array(numberOfSteps)
-        .fill(0)
-        .map((_, index) => {
-          console.log('---', this.genesisDate)
-          let date = new Date(this.genesisDate)
-          date.setSeconds(
-            date.getSeconds() + this.delay + this.cliff + this.installmentLength * index
-          )
-          let currentAmount = amount >= this.installmentWits ? this.installmentWits : amount
-          amount -= this.installmentWits
-          return {
-            date: changeDateFormat(date),
-            amount: currentAmount,
-          }
-        })
-      return steps
-    },
-    amount() {
-      return 1000
+    vesting() {
+      const vesting = calculateVesting(this.vestingInfo, this.amount, this.genesisDate)
+      this.$store.commit('setComputedVesting', vesting)
+      return vesting
     },
   },
   methods: {
